Moved actions into own package
