--- src/main/java/me/squidxtv/frameui/core/actions/ScrollAction.java
@@ -1,11 +1,0 @@
-package me.squidxtv.frameui.core.actions;
-
-import org.bukkit.entity.Player;
-import org.jetbrains.annotations.NotNull;
-
-@FunctionalInterface
-public interface ScrollAction {
-
-    void perform(@NotNull Player player, @NotNull Scrollable.ScrollDirection direction, int scrollX, int scrollY);
-
-}
