+++ src/main/java/me/squidxtv/frameui/core/actions/scroll/package-info.java
@@ -1,0 +1,2 @@
+// TODO: 19/08/2023 javascript
+package me.squidxtv.frameui.core.actions.scroll;
